Sum kernel migrated to phi
